Flip on pselect to add command console and to monitor all dlci
Monitoring the dlci channels is needed for adding support for proxying
commands between mdm6600 and w3glte so let's use pselect for that. And
as we need to decode the related AT commands for proxying, let's add a
command console too while at it. This allows typing commands for the
select DLCI using format "1 AT+CFUN?" where 1 is the DLCI followed by
a space and then the AT command to send to DLCI1.

Note that as we still don't have mdm6600 wake GPIO working, typing
commands on the console may not produce a response.

Signed-off-by: Tony Lindgren <tony@atomide.com>